chore: deduplicate sparqldecode functionality
The logic to decode SPARQL queries was implemented in the two filters
defined in this repo. To make things clearer, we remove the sparqldecode
logic from the ipmap filter and solely implement it in its own filter.
This filter needs to be added to the app-http-logger config, but it will
make the implementation of this repo a lot easier to follow.